AppDelegateIOS and AppDelegateSwift to add support for swift based projects
